spark/utils/requirements.txt
cassandra-driver==3.29.2
PyYAML==6.0.2